SymFactory: - Bugfix: Only consider cast expression as type changes if type acutally changes. - Bugfix: Treat last expression in compound braces statement as right-hand side of assignment. - Bugfix: Don't stop at address operator, see if expression is dereferenced again. - Bugfix: Search for anonymous structs in variable definitions. - Improved type change statistics.

+++ branches/cparser/insightd/structured.cpp
@@ -244,7 +244,7 @@ QString Structured::toString(QIODevice* mem, size_t offset) const
                     .arg(m->offset(), offset_len, 16, QChar('0'))
                     .arg(m->name(), -name_len)
                     .arg(m->refType()->prettyName(), -type_len)
-                    .arg(m->refTypeId(), 0, 16);
+                    .arg((uint)m->refTypeId(), 0, 16);
         }
     }
